+++ src/main/java/co/fineants/api/domain/kis/repository/infrastructure/KisAccessTokenInMemoryRepository.java
@@ -1,26 +1,26 @@
-package co.fineants.api.domain.kis.repository;
+package co.fineants.api.domain.kis.repository.infrastructure;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
 import co.fineants.api.domain.kis.client.KisAccessToken;
+import co.fineants.api.domain.kis.domain.repository.KisAccessTokenRepository;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
-@Component
-@ToString
+@Repository
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PUBLIC)
-public class KisAccessTokenInMemoryRepository {
+@ToString
+public class KisAccessTokenInMemoryRepository implements KisAccessTokenRepository {
 
 	private KisAccessToken accessToken;
 
+	@Override
 	public boolean isAccessTokenExpired(LocalDateTime dateTime) {
 		if (accessToken == null) {
 			return true;
@@ -28,10 +28,12 @@ public class KisAccessTokenInMemoryRepository {
 		return accessToken.isAccessTokenExpired(dateTime);
 	}
 
+	@Override
 	public void refreshAccessToken(KisAccessToken accessToken) {
 		this.accessToken = accessToken;
 	}
 
+	@Override
 	public String createAuthorization() {
 		if (accessToken == null) {
 			return null;
@@ -39,6 +41,7 @@ public class KisAccessTokenInMemoryRepository {
 		return accessToken.createAuthorization();
 	}
 
+	@Override
 	public boolean isTokenExpiringSoon(LocalDateTime localDateTime) {
 		if (accessToken == null) {
 			return true;
@@ -46,6 +49,7 @@ public class KisAccessTokenInMemoryRepository {
 		return accessToken.betweenSecondFrom(localDateTime).toSeconds() < 3600;
 	}
 
+	@Override
 	public Optional<KisAccessToken> getAccessToken() {
 		return Optional.ofNullable(accessToken);
 	}
