+++ src/test/java/co/fineants/api/domain/kis/repository/KisAccessTokenInMemoryRepositoryTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.test.context.ActiveProfiles;
 
 import co.fineants.api.domain.kis.client.KisAccessToken;
+import co.fineants.api.domain.kis.repository.infrastructure.KisAccessTokenInMemoryRepository;
 
 @ActiveProfiles("test")
 class KisAccessTokenInMemoryRepositoryTest {
